pyproject.toml
[tool.poetry]
name = "google_calendar_test"
version = "0.1.0"
description = ""
authors = ["mizuki sato <mizki.sato.0025@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
google-api-python-client = "^1.7"
google-auth-httplib2 = "^0.0.3"
google-auth-oauthlib = "^0.3.0"
pytz = "^2019.1"
python-dateutil = "^2.8"
iso8601 = "^0.1.12"

[tool.poetry.dev-dependencies]
pytest = "^4.6"
mypy = "^0.701.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
